Add user state for grouping conditions
Adds a user state for grouping special unique conditions
such as boosting and attacking and passes them as part of
the SM64MarioState.

Adds a field to types.h so that boosting can be checked and
modified as part of MarioState. In the future, additional items
can be added similarly to be used without defining new function
parameters.